@@ -1,9 +1,11 @@
 package main
 
 import (
-	"github.com/soulteary/flare/cmd"
+	FlareCMD "github.com/soulteary/flare/cmd"
+	FlareServer "github.com/soulteary/flare/internal/server"
 )
 
 func main() {
-	cmd.Parse()
+	flags := FlareCMD.Parse()
+	FlareServer.StartDaemon(&flags)
 }
